+++ .gitignore
@@ -353,3 +353,6 @@ tags
 [Ss]cripts
 pyvenv.cfg
 pip-selfcheck.json
+
+# Pycharm
+.idea/
