+++ README.md
@@ -1,2 +1,2 @@
-# rapid-response-datastore
+# rapid-response-xblock
 a django app plug-in for edx-platform
