--- rapid_response_datastore/__init__.py
@@ -1,2 +1,0 @@
-"""rapid_response_datastore Django app"""
-__version__ = '0.0.1'
