+++ rapid_response_xblock/__init__.py
@@ -1,0 +1,2 @@
+"""rapid_response_xblock Django app"""
+__version__ = '0.0.1'
